+++ openair2/LAYER2/nr_rlc/CMakeLists.txt
@@ -13,6 +13,7 @@ target_link_libraries(nr_rlc_core PRIVATE log_headers)
 add_library(nr_rlc STATIC
             asn1_utils.c
             nr_rlc_oai_api.c
+            nr_rlc_oai_api_nr_up.c
             nr_rlc_ue_manager.c
 )
 # L2_NR: for PDCP
