+++ openair2/LAYER2/nr_rlc/nr_rlc_entity.c
@@ -14,19 +14,6 @@
 
 #include "common/utils/time_stat.h"
 #include "common/utils/assertions.h"
-#include "openair2/LAYER2/nr_rlc/nr_rlc_oai_api_nr_up.h"
-
-static bool nr_rlc_bearer_do_drop = true;
-
-void nr_rlc_set_do_drop(bool do_drop)
-{
-  nr_rlc_bearer_do_drop = do_drop;
-}
-
-bool nr_rlc_get_do_drop(void)
-{
-  return nr_rlc_bearer_do_drop;
-}
 
 static void nr_rlc_entity_get_stats(
     nr_rlc_entity_t *entity,
